+++ fintech/src/main/java/org/weather/service/WeatherService.java
@@ -10,9 +10,9 @@ import java.util.Set;
 public interface WeatherService {
     Map<String, Double> getAverageTemperatureInRegion(Collection<Weather> weathers);
 
-    Set<String> regionFilteredByTemperature(Collection<Weather> weathers, int temperature);
+    Set<String> regionFilteredByTemperature(Collection<Weather> weathers, double temperature);
 
-    Map<Integer, List<Weather>> weatherGroupingByTemperature(Collection<Weather> weathers);
+    Map<Double, List<Weather>> weatherGroupingByTemperature(Collection<Weather> weathers);
 
-    Map<Long, List<Integer>> temperatureGroupingById(Collection<Weather> weathers);
+    Map<Long, List<Double>> temperatureGroupingById(Collection<Weather> weathers);
 }
